split stuff, condition is not runable :)
