+++ graphlearn/graphlearn.py
@@ -17,10 +17,14 @@ from sklearn.calibration import CalibratedClassifierCV
 from scipy.sparse import vstack
 from sklearn.linear_model import SGDClassifier
 
+
+from graphtools import extract_core_and_interface, core_substitution
+from feasibility import FeasibilityChecker
+from grammar import LocalSubstitutableGraphGrammar
+
 logger = logging.getLogger('log')
 logger.setLevel(logging.DEBUG)
 formatter = logging.Formatter('%(message)s')
-
 cons = logging.StreamHandler()
 cons.setLevel(logging.INFO)
 cons.setFormatter(formatter)
@@ -47,7 +51,6 @@ class GraphLearnSampler:
 
         self.same_radius = None
         self.same_core_size = None
-
         self.similarity = None
         self.sampling_interval = None
         self.n_steps = None
@@ -127,52 +130,7 @@ class GraphLearnSampler:
         self.fit_estimator(X, n_jobs)
         self.calibrate_estimator(X, nu)
 
-    ###########################  core substitution  ####################
 
-    def merge(self, G, node, node2):
-        '''
-        merge node2 into the node.
-        input nodes are strings,
-        node is the king
-        '''
-        for n in G.neighbors(node2):
-            G.add_edge(node, n)
-        G.node[node]['interface'] = True
-        G.remove_node(node2)
-
-    def find_isomorphism(self, home, other):
-        matcher = lambda x, y: x['label'] == y['label']
-        GM = iso.GraphMatcher(home, other, node_match=matcher)
-        if GM.is_isomorphic() == False:
-            return {}
-        return GM.mapping
-
-    def core_substitution(self, graph, original_cip_graph, new_cip_graph):
-        """
-        graph is the whole graph..
-        subgraph is the interfaceregrion in that we will transplant
-        new_cip_graph which is the interface and the new core
-        """
-        # select only the interfaces of the cips
-        nocore = [n for n, d in new_cip_graph.nodes(data=True) if d.has_key('core') == False]
-        newgraph_interface = nx.subgraph(new_cip_graph, nocore)
-        nocore = [n for n, d in original_cip_graph.nodes(data=True) if d.has_key('core') == False]
-        subgraph_interface = nx.subgraph(original_cip_graph, nocore)
-        # get isomorphism between interfaces, if none is found we return an empty graph
-        iso = self.find_isomorphism(subgraph_interface, newgraph_interface)
-        if len(iso) != len(subgraph_interface):
-            return nx.Graph()
-        # ok we got an isomorphism so lets do the merging
-        G = nx.union(graph, new_cip_graph, rename=('', '-'))
-        # removing old core
-        nocore = [n for n, d in original_cip_graph.nodes(data=True) if d.has_key('core')]
-        for n in nocore:
-            G.remove_node(str(n))
-        # merge interfaces
-        for k, v in iso.iteritems():
-            self.merge(G, str(k), '-' + str(v))
-        # unionizing killed my labels so we need to relabel
-        return nx.convert_node_labels_to_integers(G)
 
     ############################### SAMPLE ###########################
 
@@ -480,424 +438,3 @@ class GraphLearnSampler:
 # dont quite get it ...
 def _sample_multi(x):
     return x[1]._sample(x[0])
-
-
-################ALL THE THINGS HERE SERVE TO LEARN A GRAMMAR ############
-
-
-class core_interface_pair:
-
-    """
-    this is refered to throughout the code as cip
-    it contains the cip-graph and several pieces of information about it.
-    """
-
-    def __init__(self, ihash=0, chash=0, graph=0, radius=0, thickness=0, core_nodes_count=0, distance_dict={}):
-        self.interface_hash = ihash
-        self.core_hash = chash
-        self.graph = graph
-        self.radius = radius
-        self.thickness = thickness
-        self.core_nodes_count = core_nodes_count
-        self.distance_dict = distance_dict
-
-
-class LocalSubstitutableGraphGrammar:
-
-    """
-    the grammar.
-        can learn from graphs
-        will save all the cips in neat dictionaries
-        contains also convenience functions to make it nicely usable
-    """
-    # move all the things here that are needed to extract grammar
-
-    def __init__(self, radius_list, thickness_list, core_interface_pair_remove_threshold=3,
-                 interface_remove_threshold=2, nbit=20):
-        self.grammar = {}
-        self.interface_remove_threshold = interface_remove_threshold
-        self.radius_list = radius_list
-        self.thickness_list = thickness_list
-        self.core_interface_pair_remove_threshold = core_interface_pair_remove_threshold
-        self.vectorizer = graphlearn_utils.GraphLearnVectorizer()
-        self.hash_bitmask = 2 ** nbit - 1
-        self.nbit = nbit
-
-    def multicore_transform(self):
-        '''
-        this turns the grammar into a managed dictionary which we need for multiprocessing
-
-        note that we dont do this per default because then we cant save the grammar anymore
-        while keeping the manager outside the object
-        '''
-        # move the grammar into a manager object...
-        manager = Manager()
-        shelve = manager.dict()
-        for k, v in self.grammar.iteritems():
-            md = manager.dict()
-            for k2, v2 in v.iteritems():
-                md[k2] = v2
-            shelve[k] = md
-        self.grammar = shelve
-
-    def revert_multicore_transform(self):
-        # only if we are managed we need to do this
-        if type(self.grammar) != dict:
-            shelve = {}
-            for k, v in self.grammar.iteritems():
-                md = {}
-                for k2, v2 in v.iteritems():
-                    md[k2] = v2
-                shelve[k] = md
-            self.grammar = shelve
-
-    def difference(self, other_grammar, substract_cip_count=False):
-
-        # i assume set() my even be faster than sort...
-        # should be even faster to just test -_-
-
-        # mykeys=set(my.keys())
-        # otherkeys=set(other_grammar.keys())
-        for ihash in self.grammar.keys():
-            if ihash in other_grammar:
-                # so if my ihash is in both i can compare the corehashes
-                for chash in self.grammar[ihash].keys():
-                    if chash in other_grammar[ihash]:
-                        if substract_cip_count:
-                            self.grammar[ihash][chash].count -= self.grammar[ihash][chash].count
-                        else:
-                            self.grammar[ihash].pop(chash)
-        if substract_cip_count:
-            self.clean()
-
-    def union(self, other_grammar, cip_count_function=sum):
-        '''
-        we build the union of grammars..
-        '''
-
-        for ihash in other_grammar.keys():
-            if ihash in self.grammar:
-                # ok so we are in the intersection...
-                for chash in other_grammar[ihash]:
-                    if chash in self.grammar[ihash]:
-
-                        self.grammar[ihash][chash].counter = cip_count_function(
-                            self.grammar[ihash][chash].counter, other_grammar[ihash][chash].counter)
-                    else:
-                        self.grammar[ihash][chash] = other_grammar[ihash][chash]
-            else:
-                self.grammar[ihash] = other_grammar[ihash]
-
-    def intersect(self, other_grammar):
-        # counts will be added...
-
-        for ihash in self.grammar.keys():
-            if ihash in other_grammar:
-                # so if self.grammar ihash is in both i can compare the corehashes
-                for chash in self.grammar[ihash].keys():
-                    if chash in other_grammar[ihash]:
-                        self.grammar[ihash][chash].counter = min(self.grammar[ihash][chash].counter, other_grammar[ihash][chash].counter)
-                    else:
-                        self.grammar[ihash].pop(chash)
-            else:
-                self.grammar.pop(ihash)
-
-    def clean(self):
-        """
-            after adding many cips to the grammar it is possible
-            that some thresholds are not reached, and we dont want unnecessary ballast in
-            our grammar.
-            so we clean up.
-        """
-        for interface in self.grammar.keys():
-
-            for core in self.grammar[interface].keys():
-                if self.grammar[interface][core].count < self.core_interface_pair_remove_threshold:
-                    self.grammar[interface].pop(core)
-
-            if len(self.grammar[interface]) < self.interface_remove_threshold:
-                self.grammar.pop(interface)
-
-    def add_same_radius_quicklookup(self):
-        '''
-            there is now self.radiuslookup{ interfacehash: {radius:[list of corehashes with that ihash and radius]} }
-        '''
-        self.radiuslookup = {}
-        for interface in self.grammar.keys():
-            radius_lookup = [[]] * (max(self.radius_list) + 1)
-
-            for core in self.grammar[interface].keys():
-                radius = self.grammar[interface][core].radius
-                if radius in radius_lookup:
-                    radius_lookup[radius].append(core)
-                else:
-                    radius_lookup[radius] = [core]
-            self.radiuslookup[interface] = radius_lookup
-
-    def add_core_size_quicklookup(self):
-        '''
-            there is now self.radiuslookup{ interfacehash: {radius:[list of corehashes with that ihash and radius]} }
-        '''
-        self.core_size = {}
-        for interface in self.grammar.keys():
-            core_size = {}
-            for core in self.grammar[interface].keys():
-                nodes_count = self.grammar[interface][core].core_nodes_count
-                if nodes_count in core_size:
-                    core_size[nodes_count].append(core)
-                else:
-                    core_size[nodes_count] = [core]
-            self.core_size[interface] = core_size
-
-    def read(self, graphs, n_jobs=-1):
-        '''
-        we extract all chips from graphs of a graph iterator
-        we use n_jobs processes to do so.
-        '''
-
-        # if we should use only one process we use read_single ,  else read_multi
-        if n_jobs == 1:
-            self.read_single(graphs)
-        else:
-            self.read_multi(graphs, n_jobs)
-
-    def grammar_add_core_interface_data(self, cid):
-        '''
-            cid is a core interface data instance.
-            we will add the cid to our grammar.
-        '''
-
-        # initialize gramar[interfacehash] if necessary
-        if cid.interface_hash not in self.grammar:
-            self.grammar[cid.interface_hash] = {}
-
-        # initialize or get grammar[interfacehash][corehash] which is now called subgraph_data
-        if cid.core_hash in self.grammar[cid.interface_hash]:
-            subgraph_data = self.grammar[cid.interface_hash][cid.core_hash]
-        else:
-            subgraph_data = core_interface_pair()
-            self.grammar[cid.interface_hash][cid.core_hash] = subgraph_data
-            subgraph_data.count = 0
-
-        # put new information in the subgraph_data
-        # we only save the count until we know that we will keep the actual cip
-        subgraph_data.count += 1
-        if subgraph_data.count == self.core_interface_pair_remove_threshold:
-            subgraph_data.__dict__.update(cid.__dict__)
-            subgraph_data.count = self.core_interface_pair_remove_threshold
-
-    def read_single(self, graphs):
-        """
-            for graph in graphs:
-                get cips of graph
-                    put cips into grammar
-        """
-        for gr in graphs:
-            problem = (gr, self.radius_list, self.thickness_list, self.vectorizer, self.hash_bitmask)
-            for core_interface_data_list in extract_cores_and_interfaces(problem):
-                for cid in core_interface_data_list:
-                    self.grammar_add_core_interface_data(cid)
-
-    def read_multi(self, graphs, n_jobs):
-        """
-        will take graphs and to multiprocessing to extract their cips
-        and put these cips in the grammar
-        """
-
-        # generate iterator of problem instances
-        problems = itertools.izip(graphs, itertools.repeat(self.radius_list),
-                                  itertools.repeat(self.thickness_list),
-                                  itertools.repeat(self.vectorizer),
-                                  itertools.repeat(self.hash_bitmask))
-
-        # creating pool of workers
-        if n_jobs == -1:
-            pool = Pool()
-        else:
-            pool = Pool(processes=n_jobs)
-        # distributing jobs to workers
-        result = pool.imap_unordered(extract_cores_and_interfaces, problems, 10)
-        # the resulting chips can now be put intro the grammar
-        for core_interface_data_listlist in result:
-            for core_interface_data_list in core_interface_data_listlist:
-                for cid in core_interface_data_list:
-                    self.grammar_add_core_interface_data(cid)
-
-        pool.close()
-
-
-def extract_cores_and_interfaces(parameters):
-    # unpack arguments, expand the graph
-    graph, radius_list, thickness_list, vectorizer, hash_bitmask = parameters
-    graph = graphlearn_utils.expand_edges(graph)
-    cips = []
-    for node in graph.nodes_iter():
-        if 'edge' in graph.node[node]:
-            continue
-        core_interface_list = extract_core_and_interface(node, graph, radius_list, thickness_list,
-                                                         vectorizer=vectorizer, hash_bitmask=hash_bitmask)
-        if core_interface_list:
-            cips.append(core_interface_list)
-    return cips
-
-
-# ############################### FEASIBILITY CHECKER ###################
-
-
-class FeasibilityChecker():
-
-    def __init__(self):
-        self.checklist = []
-        self.checklist.append(defaultcheck)
-
-    def check(self, ng):
-        for f in self.checklist:
-            if f(ng) == False:
-
-                return False
-        return True
-
-
-def defaultcheck(ng):
-    if len(ng) < 1:
-        logger.debug('graph non existent')
-        return False
-    for node_id in ng.nodes_iter():
-        if 'edge' in ng.node[node_id]:
-            if len(ng.neighbors(node_id)) != 2:
-                logger.debug('feasibility- edge check failed, (interface twist phenomenon probably)')
-                return False
-    return True
-
-
-#####################################   extract a core/interface pair #####################
-
-
-def invert_dict(d):
-    """
-    so input is usualy a distance dictionaty so
-    {nodenumber: distance, nodenumber:distance} we turn this into {distance: [nodenumber, nodenumber]}
-    """
-    d2 = {}
-    for k, v in d.iteritems():
-        l = []
-        d2[v] = d2.get(v, l)
-        d2[v].append(k)
-    return d2
-
-
-def calc_interface_hash(interface_graph, hash_bitmask):
-    """
-        so we calculate a hash of a graph
-    """
-    l = []
-    node_name_cache = {}
-
-    all_nodes = set(interface_graph.nodes())
-    visited = set()
-    # all the edges
-    for (a, b) in interface_graph.edges():
-        visited.add(a)
-        visited.add(b)
-
-        ha = node_name_cache.get(a, -1)
-        if ha == -1:
-            ha = calc_node_name(interface_graph, a, hash_bitmask)
-            node_name_cache[a] = ha
-        hb = node_name_cache.get(b, -1)
-        if hb == -1:
-            hb = calc_node_name(interface_graph, b, hash_bitmask)
-            node_name_cache[b] = hb
-        l.append((ha ^ hb) + (ha + hb))
-    l.sort()
-
-    # nodes that dont have edges
-    l += [interface_graph.node[node_id]['hlabel'][0] for node_id in all_nodes - visited]
-    l = fast_hash(l, hash_bitmask)
-    return l
-
-
-def calc_core_hash(core_graph, hash_bitmask):
-    return calc_interface_hash(core_graph, hash_bitmask)
-
-
-def calc_node_name(interfacegraph, node, hash_bitmask):
-    '''
-     part of generating the hash for a graph is calculating the hash of a node in the graph
-    '''
-    d = nx.single_source_shortest_path_length(interfacegraph, node, 20)
-    # d is now node:dist
-    # l is a list of  hash(label,distance)
-    #l=[   func([interfacegraph.node[nid]['intlabel'],dis])  for nid,dis in d.items()]
-    l = [interfacegraph.node[nid]['hlabel'][0] + dis for nid, dis in d.items()]
-    l.sort()
-    l = fast_hash(l, hash_bitmask)
-    return l
-
-
-def extract_core_and_interface(node, graph, radius_list=None, thickness_list=None, vectorizer=None,
-                               hash_bitmask=2 * 20 - 1):
-    """
-
-:param node: root node
-:param graph: graph
-:param radius_list:
-:param thickness_list:
-:param vectorizer: a vectorizer
-:param hash_bitmask:
-
-
-:return: radius_list*thicknes_list long list of cips
-"""
-
-    if 'hlabel' not in graph.node[0]:
-        vectorizer._label_preprocessing(graph)
-
-    # which nodes are in the relevant radius
-    dist = nx.single_source_shortest_path_length(graph, node, max(radius_list) + max(thickness_list))
-    # we want the relevant subgraph and we want to work on a copy
-    retgraph = nx.Graph(graph.subgraph(dist))
-
-    # we want to inverse the dictionary.
-    # so now we see {distance:[list of nodes at that distance]}
-    nodedict = invert_dict(dist)
-
-    cip_list = []
-    for thickness_ in thickness_list:
-        for radius_ in radius_list:
-
-            # see if it is feasable to extract
-            if radius_ + thickness_ not in nodedict:
-                continue
-
-            # calculate hashes
-            # d={1:[1,2,3],2:[3,4,5]}
-            # print [ i for x in [1,2] for i in d[x] ]
-            interface_graph_nodes = [item for x in range(radius_ + 1, radius_ + thickness_ + 1) for item in
-                                     nodedict.get(x, [])]
-            interfacehash = calc_interface_hash(retgraph.subgraph(interface_graph_nodes), hash_bitmask)
-
-            core_graph_nodes = [item for x in range(radius_ + 1) for item in nodedict.get(x, [])]
-            corehash = calc_core_hash(retgraph.subgraph(core_graph_nodes), hash_bitmask)
-
-            # get relevant subgraph
-            nodes = [node for i in range(radius_ + thickness_ + 1) for node in nodedict[i]]
-            cip_graph = nx.Graph(retgraph.subgraph(nodes))
-
-            # marking cores and interfaces in subgraphs
-            for i in range(radius_ + 1):
-                for no in nodedict[i]:
-                    cip_graph.node[no]['core'] = True
-                    if 'interface' in cip_graph.node[no]:
-                        cip_graph.node[no].pop('interface')
-            for i in range(radius_ + 1, radius_ + thickness_ + 1):
-                if i in nodedict:
-                    for no in nodedict[i]:
-                        cip_graph.node[no]['interface'] = True
-                        if 'core' in cip_graph.node[no]:
-                            cip_graph.node[no].pop('core')
-
-            core_nodes_count = sum([len(nodedict[x]) for x in range(radius_ + 1)])
-
-            cip_list.append(core_interface_pair(interfacehash, corehash, cip_graph, radius_, thickness_, core_nodes_count, distance_dict=nodedict))
-    return cip_list
